1. 修复读取Excel公式单元格值，第一次公式不能计算的BUG 2. POI 版本升级4.1.2   ---->   5.1.0 3. commons-lang3版本升级3.3.2  ---->  3.12.0 4. 错误日志输出

+++ src/test/java/excel/imports/data/DemoData.java
@@ -1,12 +1,11 @@
 package excel.imports.data;
 
-import java.math.BigDecimal;
 import java.util.Date;
 import java.util.Map;
 
 public class DemoData {
 
-	private BigDecimal bigDecimalHandler;
+	private Long bigDecimalHandler;
 
 	private Boolean booleanHandler;
 
@@ -38,11 +37,11 @@ public class DemoData {
 		this.shortHandler = shortHandler;
 	}
 
-	public BigDecimal getBigDecimalHandler() {
+	public Long getBigDecimalHandler() {
 		return bigDecimalHandler;
 	}
 
-	public void setBigDecimalHandler(BigDecimal bigDecimalHandler) {
+	public void setBigDecimalHandler(Long bigDecimalHandler) {
 		this.bigDecimalHandler = bigDecimalHandler;
 	}
